producers/user_data_wh/requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Faker==14.0.0
idna==3.3
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-pkcs12==1.14
six==1.16.0
urllib3==1.26.11
